Rewrite package as a commonjs module and remove esm dependency
