+++ helpers/exec.js
@@ -1,4 +1,4 @@
-import { exec } from 'child_process';
-import { promisify } from 'util';
+const { exec } = require('child_process');
+const { promisify } = require('util');
 
-export default promisify(exec);
+module.exports.default = promisify(exec);
