+++ helpers/prompts.js
@@ -1,8 +1,8 @@
-import p from 'prompts';
+const prompts = require('prompts');
 
-export const exit = (code = 0) => {
+module.exports.exit = (code = 0) => {
   console.log(`\n${code ? '😵 Exiting.' : '✨ Done.'} Thanks for using git-add-account!\n`);
   process.exit(code);
 };
 
-export default (prompts, options) => p(prompts, { onCancel: () => exit(1), ...options });
+module.exports.default = (questions, options) => prompts(questions, { onCancel: () => exit(1), ...options });
