+++ helpers/validate.js
@@ -1,6 +1,8 @@
-export { z } from 'zod';
+const { z } = require('zod');
 
-export default (schema) => (value) => {
+module.exports.z = z;
+
+module.exports.default = (schema) => (value) => {
   const { success, error } = schema.safeParse(value);
   if (!success) return error.format()._errors[0];
   return true;
