+++ package.json
@@ -25,17 +25,16 @@
     "type": "git",
     "url": "git+https://github.com/savaryna/git-add-account.git"
   },
+  "type": "commonjs",
   "bin": {
     "gac": "index.js",
     "git-add-account": "index.js"
   },
   "main": "index.js",
-  "module": "main.js",
   "os": [
     "darwin"
   ],
   "dependencies": {
-    "esm": "^3.2.25",
     "prompts": "^2.4.2",
     "zod": "^3.20.6"
   },
